Correction d'une stupidité.

+++ Scripts/clean.sh
@@ -7,18 +7,22 @@ tmpFold="Temp"
 binFold_i="Bin"
 outFold_i="Out"
 
-status="non"
-until [ ${status} = "oui" ]
-do
+echo ${binFold}":"
+ls ${binFold}
+echo ${outFold}":"
+ls ${outFold}
+echo ${tmpFold}":" 
+ls ${tmpFold}
 
-	echo ${binFold} ${outFold} ${tmpFold}
-	echo "Etes-vous sûr de vouloir effacer les fichiers ? (oui/non)"
-	read status
+echo "Etes-vous sûr de vouloir effacer les fichiers ? (oui/non)"
+read status
+	
+if [ ${status} = "oui" ]
+then
 
 	rm -rf ${binFold}/${binFold_i}*
 	rm -rf ${outFold}/${outFold_i}*
 	rm -rf ${tmpFold}/*.out
-
-done
-
-	
+else
+	echo "Rien n'a été effacé."
+fi
